mabatis selectkey 를 이용해서 실행후 값 넣어주기..

+++ SpringWork/HelloSpring/src/main/java/com/dm/spring/board/model/vo/Attachment.java
@@ -14,7 +14,7 @@ import lombok.NoArgsConstructor;
 @Builder
 public class Attachment {
 	private long attachmentNo;
-	private int boardNo;
+	private long boardNo;
 	private String originalFileName;
 	private String renamedFileName;
 	private Date uploadDate;
